+++ jscomp/test/gpr_3519_test.ml
@@ -9,15 +9,15 @@ type 'a arra = 'a array
 
 external
   f0 :
-  int -> int -> int array -> unit
+  int -> int -> int array -> (int [@mel.this]) -> unit
   = "f0"
-  [@@mel.send.pipe:int]
+  [@@mel.send]
   [@@mel.variadic]
 
 external
   f1 :
-  int -> int -> y:int array -> unit
+  int -> int -> y:int array -> (int [@mel.this]) -> unit
   = "f1"
-  [@@mel.send.pipe:int]
+  [@@mel.send]
   [@@mel.variadic]